refactor: move common services to libs
