+++ src/MakaMek.Avalonia/MakaMek.Avalonia/MakaMek.Avalonia.csproj
@@ -19,7 +19,6 @@
 
     <ItemGroup>
         <PackageReference Include="Avalonia" Version="12.0.4" />
-        <PackageReference Include="ReactiveUI.Avalonia" Version="14.7.1" />
         <PackageReference Include="Avalonia.Skia" Version="12.0.4" />
         <PackageReference Include="Avalonia.Themes.Fluent" Version="12.0.4" />
         <PackageReference Include="Avalonia.Fonts.Inter" Version="12.0.4" />
