+++ src/MakaMek.Avalonia/MakaMek.Services.Avalonia/MakaMek.Services.Avalonia.csproj
@@ -15,6 +15,7 @@
 
     <ItemGroup>
         <PackageReference Include="Avalonia" Version="12.0.4" />
+        <PackageReference Include="ReactiveUI.Avalonia" Version="12.0.1" />
     </ItemGroup>
 
     <ItemGroup>
