+++ src/MakaMek.Services/MakaMek.Services.csproj
@@ -14,6 +14,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.8" />
       <PackageReference Include="PdfSharpCore" Version="1.3.67" />
     </ItemGroup>
