+++ src/MakaMek.Services/ResourceProviders/GitHubResourceStreamProvider.cs
@@ -2,9 +2,8 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Microsoft.Extensions.Logging;
-using Sanet.MakaMek.Services;
 
-namespace Sanet.MakaMek.Core.Services.ResourceProviders;
+namespace Sanet.MakaMek.Services.ResourceProviders;
 
 public class GitHubResourceStreamProvider : IResourceStreamProvider
 {
