+++ src/MakaMek.Services/ResourceProviders/IResourceStreamProvider.cs
@@ -1,4 +1,4 @@
-namespace Sanet.MakaMek.Core.Services;
+namespace Sanet.MakaMek.Services.ResourceProviders;
 
 /// <summary>
 /// Interface for providing resource streams from various sources (assemblies, filesystem, network, etc.)
