+++ src/MakaMek.Tools/BotContainer/DependencyInjection/BotContainerServices.cs
@@ -11,11 +11,11 @@ using Sanet.MakaMek.Core.Services;
 using Sanet.MakaMek.Core.Services.Cryptography;
 using Sanet.MakaMek.Services;
 using Sanet.MakaMek.Core.Services.Logging.Factories;
-using Sanet.MakaMek.Core.Services.ResourceProviders;
 using Sanet.MakaMek.Core.Services.Transport;
 using Sanet.MakaMek.Core.Utils;
 using Sanet.MakaMek.Localization;
 using Sanet.MakaMek.Map.Factories;
+using Sanet.MakaMek.Services.ResourceProviders;
 using Sanet.MakaMek.Tools.BotContainer.Configuration;
 using Sanet.MakaMek.Tools.BotContainer.Models.Mcp.Tools;
 using Sanet.MakaMek.Tools.BotContainer.Services;
