+++ src/MakaMek.Tools/BotContainer/Services/HeadlessDispatcherService.cs
@@ -1,5 +1,6 @@
 using System.Reactive.Concurrency;
 using Sanet.MakaMek.Core.Services;
+using Sanet.MakaMek.Services;
 
 namespace Sanet.MakaMek.Tools.BotContainer.Services;
